continue splitting up
