+++ src/generators/members/index.ts
@@ -1,3 +1,5 @@
 export * from "./server-constructor.js";
-export * from "./server-methods.js";
+export * from "./server-handle-method.js";
 export * from "./server-properties.js";
+export * from "./server-register-methods.js";
+export * from "./server-route-handle-methods.js";
