+++ src/generators/members/server-route-handle-methods.ts
@@ -3,171 +3,18 @@ import * as models from "../../models/index.js";
 import { toCamel, toPascal } from "../../utils/name.js";
 import { CodeGeneratorBase } from "../code-generator-base.js";
 
-export class ServerMethodsCodeGenerator extends CodeGeneratorBase {
+export class ServerRouteHandleMethodsCodeGenerator extends CodeGeneratorBase {
   public *getStatements() {
-    yield* this.generateHandleMethodDeclarations();
-    yield* this.generateRegisterOperationMethodsDeclarations();
-    yield* this.generateRegisterAuthorizationMethodsDeclarations();
-    yield* this.generateRouteHandlersMethodsDeclarations();
+    yield* this.generateRouteHandlerMethodsDeclarations();
   }
-
-  //#region handle
-
-  /**
-   * generate handler for incoming requests
-   */
-  private *generateHandleMethodDeclarations() {
-    const { factory: f } = this;
-
-    yield f.createMethodDeclaration(
-      [f.createToken(ts.SyntaxKind.PublicKeyword)],
-      undefined,
-      "handle",
-      undefined,
-      undefined,
-      [
-        f.createParameterDeclaration(
-          undefined,
-          undefined,
-          "incomingRequest",
-          undefined,
-          f.createTypeReferenceNode(
-            f.createQualifiedName(
-              f.createIdentifier("lib"),
-              f.createIdentifier("ServerIncomingRequest"),
-            ),
-          ),
-        ),
-      ],
-      f.createTypeReferenceNode(
-        f.createQualifiedName(
-          f.createIdentifier("lib"),
-          f.createIdentifier("ServerOutgoingResponse"),
-        ),
-        undefined,
-      ),
-      f.createBlock([...this.generateHandlerFunctionStatements()], true),
-    );
-  }
-  private *generateHandlerFunctionStatements() {
-    const { factory: f } = this;
-
-    yield f.createVariableStatement(
-      undefined,
-      f.createVariableDeclarationList(
-        [
-          f.createVariableDeclaration(
-            f.createArrayBindingPattern([
-              f.createBindingElement(
-                undefined,
-                undefined,
-                f.createIdentifier("routeKey"),
-                undefined,
-              ),
-              f.createBindingElement(
-                undefined,
-                undefined,
-                f.createIdentifier("routeParameters"),
-                undefined,
-              ),
-            ]),
-            undefined,
-            undefined,
-            f.createCallExpression(
-              f.createPropertyAccessExpression(
-                f.createPropertyAccessExpression(
-                  f.createThis(),
-                  f.createIdentifier("router"),
-                ),
-                f.createIdentifier("parseRoute"),
-              ),
-              undefined,
-              [
-                f.createPropertyAccessExpression(
-                  f.createIdentifier("incomingRequest"),
-                  f.createIdentifier("path"),
-                ),
-              ],
-            ),
-          ),
-        ],
-        ts.NodeFlags.Const,
-      ),
-    );
-
-    yield f.createSwitchStatement(
-      f.createIdentifier("routeKey"),
-      f.createCaseBlock([...this.generateHandleMethodPathCaseClauses()]),
-    );
-  }
-  private *generateHandleMethodPathCaseClauses() {
-    const { factory: f } = this;
-
-    for (
-      let pathIndex = 0;
-      pathIndex < this.apiModel.paths.length;
-      pathIndex++
-    ) {
-      const pathModel = this.apiModel.paths[pathIndex];
-      yield f.createCaseClause(f.createNumericLiteral(pathIndex + 1), [
-        f.createSwitchStatement(
-          f.createPropertyAccessExpression(
-            f.createIdentifier("incomingRequest"),
-            f.createIdentifier("method"),
-          ),
-          f.createCaseBlock([
-            ...this.generateHandleMethodOperationCaseClauses(pathModel),
-          ]),
-        ),
-      ]);
-    }
-
-    yield f.createDefaultClause([
-      f.createThrowStatement(f.createStringLiteral("not found")),
-    ]);
-  }
-  private *generateHandleMethodOperationCaseClauses(pathModel: models.Path) {
-    const { factory: f } = this;
-
-    for (const operationModel of pathModel.operations) {
-      const routeHandlerName = toCamel("handle", operationModel.name, "route");
-
-      yield f.createCaseClause(
-        f.createStringLiteral(operationModel.method.toUpperCase()),
-        [
-          f.createReturnStatement(
-            f.createCallExpression(
-              f.createPropertyAccessExpression(
-                f.createThis(),
-                routeHandlerName,
-              ),
-              undefined,
-              [
-                f.createIdentifier("routeParameters"),
-                f.createIdentifier("incomingRequest"),
-              ],
-            ),
-          ),
-        ],
-      );
-    }
-
-    yield f.createDefaultClause([
-      f.createThrowStatement(f.createStringLiteral("method not supported")),
-    ]);
-  }
-
-  //#endregion
-
-  //#region routes
 
   /**
    * all route handler functions
    */
-  private *generateRouteHandlersMethodsDeclarations() {
+  private *generateRouteHandlerMethodsDeclarations() {
     for (const pathModel of this.apiModel.paths) {
       for (const operationModel of pathModel.operations) {
-        yield* this.generateRouteHandlersFunctionDeclarations(
+        yield* this.generateRouteHandlerMethodDeclarations(
           pathModel,
           operationModel,
         );
@@ -180,7 +27,7 @@ export class ServerMethodsCodeGenerator extends CodeGeneratorBase {
    * @param pathModel
    * @param operationModel
    */
-  private *generateRouteHandlersFunctionDeclarations(
+  private *generateRouteHandlerMethodDeclarations(
     pathModel: models.Path,
     operationModel: models.Operation,
   ) {
@@ -227,7 +74,7 @@ export class ServerMethodsCodeGenerator extends CodeGeneratorBase {
       ),
       f.createBlock(
         [
-          ...this.generateRouteHandlersFunctionStatements(
+          ...this.generateRouteHandlersMethodStatements(
             pathModel,
             operationModel,
           ),
@@ -242,7 +89,7 @@ export class ServerMethodsCodeGenerator extends CodeGeneratorBase {
    * @param pathModel
    * @param operationModel
    */
-  private *generateRouteHandlersFunctionStatements(
+  private *generateRouteHandlersMethodStatements(
     pathModel: models.Path,
     operationModel: models.Operation,
   ) {
@@ -582,174 +429,4 @@ export class ServerMethodsCodeGenerator extends CodeGeneratorBase {
 
     yield f.createThrowStatement(f.createStringLiteral("TODO"));
   }
-
-  //#endregion
-
-  //#region operations
-
-  /**
-   * register functions for all operation handlers
-   */
-  private *generateRegisterOperationMethodsDeclarations() {
-    for (const pathModel of this.apiModel.paths) {
-      for (const operationModel of pathModel.operations) {
-        yield* this.generateRegisterOperationMethodDeclarations(
-          pathModel,
-          operationModel,
-        );
-      }
-    }
-  }
-
-  /**
-   * register functions for a single operation handler
-   * @param pathModel
-   * @param operationModel
-   */
-  private *generateRegisterOperationMethodDeclarations(
-    pathModel: models.Path,
-    operationModel: models.Operation,
-  ) {
-    const { factory: f } = this;
-    const methodName = toCamel("register", operationModel.name, "operation");
-    const operationHandlerTypeName = toPascal(
-      operationModel.name,
-      "operation",
-      "handler",
-    );
-
-    yield f.createMethodDeclaration(
-      [f.createToken(ts.SyntaxKind.PublicKeyword)],
-      undefined,
-      methodName,
-      undefined,
-      undefined,
-      [
-        f.createParameterDeclaration(
-          undefined,
-          undefined,
-          "operationHandler",
-          undefined,
-          f.createTypeReferenceNode(operationHandlerTypeName),
-        ),
-      ],
-      undefined,
-      f.createBlock(
-        [
-          ...this.generateRegisterOperationMethodStatements(
-            pathModel,
-            operationModel,
-          ),
-        ],
-        true,
-      ),
-    );
-  }
-
-  /**
-   * statements for registering an operation handler
-   * @param pathModel
-   * @param operationModel
-   */
-  private *generateRegisterOperationMethodStatements(
-    pathModel: models.Path,
-    operationModel: models.Operation,
-  ) {
-    const { factory: f } = this;
-
-    const operationHandlerName = toCamel(
-      "handle",
-      operationModel.name,
-      "operation",
-    );
-
-    yield f.createExpressionStatement(
-      f.createBinaryExpression(
-        f.createPropertyAccessExpression(
-          f.createThis(),
-          f.createIdentifier(operationHandlerName),
-        ),
-        f.createToken(ts.SyntaxKind.EqualsToken),
-        f.createIdentifier("operationHandler"),
-      ),
-    );
-  }
-
-  //#endregion
-
-  //#region authorization
-
-  private *generateRegisterAuthorizationMethodsDeclarations() {
-    for (const authorizationModel of this.apiModel.authorizations) {
-      yield* this.generateRegisterAuthorizationMethodDeclarations(
-        authorizationModel,
-      );
-    }
-  }
-
-  private *generateRegisterAuthorizationMethodDeclarations(
-    authorizationModel: models.Authorization,
-  ) {
-    const { factory: f } = this;
-    const methodName = toCamel(
-      "register",
-      authorizationModel.name,
-      "authorization",
-    );
-
-    yield f.createMethodDeclaration(
-      [f.createToken(ts.SyntaxKind.PublicKeyword)],
-      undefined,
-      methodName,
-      undefined,
-      undefined,
-      [
-        f.createParameterDeclaration(
-          undefined,
-          undefined,
-          "authorizationHandler",
-          undefined,
-          f.createFunctionTypeNode(
-            undefined,
-            [],
-            f.createKeywordTypeNode(ts.SyntaxKind.VoidKeyword),
-          ),
-        ),
-      ],
-      undefined,
-      f.createBlock(
-        [
-          ...this.generateRegisterAuthorizationMethodStatements(
-            authorizationModel,
-          ),
-        ],
-        true,
-      ),
-    );
-  }
-
-  private *generateRegisterAuthorizationMethodStatements(
-    authorizationModel: models.Authorization,
-  ) {
-    const { factory: f } = this;
-
-    const operationHandlerName = toCamel(
-      "handle",
-      authorizationModel.name,
-      "authorization",
-    );
-
-    yield f.createExpressionStatement(
-      f.createBinaryExpression(
-        f.createPropertyAccessExpression(
-          f.createThis(),
-          f.createIdentifier(operationHandlerName),
-        ),
-        f.createToken(ts.SyntaxKind.EqualsToken),
-        f.createIdentifier("authorizationHandler"),
-      ),
-    );
-  }
-
-  //#endregion
 }
