+++ src/generators/types/server.ts
@@ -2,9 +2,11 @@ import ts from "typescript";
 import { CodeGeneratorBase } from "../code-generator-base.js";
 import {
   ServerConstructorCodeGenerator,
-  ServerMethodsCodeGenerator,
   ServerPropertiesCodeGenerator,
+  ServerRegisterMethodsCodeGenerator,
+  ServerRouteHandleMethodsCodeGenerator,
 } from "../members/index.js";
+import { ServerHandleMethodCodeGenerator } from "../members/server-handle-method.js";
 
 export class ServerTypeCodeGenerator extends CodeGeneratorBase {
   private serverConstructorCodeGenerator = new ServerConstructorCodeGenerator(
@@ -15,7 +17,11 @@ export class ServerTypeCodeGenerator extends CodeGeneratorBase {
     this.factory,
     this.apiModel,
   );
-  private serverMembersCodeGenerator = new ServerMethodsCodeGenerator(
+  private serverRegisterMethodsCodeGenerator =
+    new ServerRegisterMethodsCodeGenerator(this.factory, this.apiModel);
+  private serverRouteHandleMethodsCodeGenerator =
+    new ServerRouteHandleMethodsCodeGenerator(this.factory, this.apiModel);
+  private serverHandleMethodCodeGenerator = new ServerHandleMethodCodeGenerator(
     this.factory,
     this.apiModel,
   );
@@ -62,6 +68,8 @@ export class ServerTypeCodeGenerator extends CodeGeneratorBase {
   private *generateServerElementsDeclarations() {
     yield* this.serverPropertiesCodeGenerator.getStatements();
     yield* this.serverConstructorCodeGenerator.getStatements();
-    yield* this.serverMembersCodeGenerator.getStatements();
+    yield* this.serverRegisterMethodsCodeGenerator.getStatements();
+    yield* this.serverHandleMethodCodeGenerator.getStatements();
+    yield* this.serverRouteHandleMethodsCodeGenerator.getStatements();
   }
 }
